Review cS and rename Signature/Verify
