+++ src/Extract.php
@@ -12,16 +12,6 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-/*
- * This file is part of the box project.
- *
- * (c) Kevin Herrera <kevin@herrera.io>
- *     Théo Fidry <theo.fidry@gmail.com>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace KevinGH\Box;
 
 use InvalidArgumentException;
