+++ src/Verifier.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Signature;
+namespace KevinGH\Box;
 
 use KevinGH\Box\Exception\SignatureException;
