+++ src/Verifier/BufferedHash.php
@@ -12,7 +12,9 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Signature;
+namespace KevinGH\Box\Verifier;
+
+use KevinGH\Box\Verifier;
 
 /**
  * Buffers the hash as opposed to updating incrementally.
