+++ src/Verifier/OpenSsl.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Signature;
+namespace KevinGH\Box\Verifier;
 
 use KevinGH\Box\Exception\OpenSslExceptionFactory;
 use RuntimeException;
