+++ tests/Verifier/PhpSecLibTest.php
@@ -12,22 +12,22 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Signature;
+namespace KevinGH\Box\Verifier;
 
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \KevinGH\Box\Signature\PublicKeyDelegate
+ * @covers \KevinGH\Box\Verifier\PhpSecLib
  */
-class PublicKeyDelegateTest extends TestCase
+class PhpSecLibTest extends TestCase
 {
     private const FIXTURES_DIR = __DIR__.'/../../fixtures/signed_phars';
 
-    public function testFunctional(): void
+    public function test_it_can_verify_a_file(): void
     {
         $path = self::FIXTURES_DIR.'/openssl.phar';
 
-        $hash = new PublicKeyDelegate('openssl', $path);
+        $hash = new PhpSecLib('openssl', $path);
 
         $hash->update(
             file_get_contents($path, false, null, 0, filesize($path) - 76)
